Retain info about the keyboard definition even when current device doesn't support it
In order to show the user info about *which* keyboard is not supported, and in the future allow them to request it, we need to know what keyboard we're talking about. Using a failable initializer was safer than this, but kept us from knowing anything about the keyboard definition in question. Will probably refactor to make `layoutForCurrentDevice` an optional property of `KeyboardDefinition` so this is safer and more clear.