Enhance missing layout message (#237)
* Get building on Xcode 16.1

* Add globe button to switch keyboards when device not supported

* Fix for crash when switching between light/dark mode

I don't like it

* Revert "Fix for crash when switching between light/dark mode"

This reverts commit 7e0a7d2b3c0f535559119cd449f2f5e2d25021ac.

* Better fix crash when switching between dark/light mode

Still don't love it but it's much fewer code changes and should be safe enough

* Localize keyboard not supported message

* Retain info about the keyboard definition even when current device doesn't support it

In order to show the user info about *which* keyboard is not supported, and in the future allow them to request it, we need to know what keyboard we're talking about. Using a failable initializer was safer than this, but kept us from knowing anything about the keyboard definition in question. Will probably refactor to make `layoutForCurrentDevice` an optional property of `KeyboardDefinition` so this is safer and more clear.

* Add WIP email button that drafts keyboard request email in Mail.app

* Make helper class for opening URLs

We can't have nice things from an App Extension. This workaround lets us open URLs despite not having access to UIApplication.shared.

* Remove unused method

* Remove todo

* Move emailButtonTapped method

* Make UI decent, update copy, dummy github issue button

* Request keyboard via github action, pretty print device name

* Revert "Get building on Xcode 16.1"

This reverts commit 8c0bdb589acb991c4bd37b5cb90f62c459321186.

* Revert to non-optional keyboard mode definitions

Test for empty instead of nil