I hate that this is the fix for duplicate strings upon export to Pontoon
Turns out exporting strings will always add any discovered NSLocalizedStrings in code to the BaseKeyboard lproj, even if those strings are commented out in code, and even if a BaseKeyboard lproj doesn't exist. Since the Localizable.strings file is added to the Keyboard target, this means the export finds these strings twice: once for Keyboard and once for BaseKeyboard. The workaround is to add Localizable.strings to the BaseKeyboard target as well - then it considers the strings to only be a part of Keyboard, which results in no more duplicate strings showing up in Pontoon.