Pontoon localization (#241)
* Get building in Xcode 16.2

* Copy locales settings from ios-pontoon-test

* Add exported strings for all locales

* Fix nb naming

* Move locales to different repo

* Run initial import from l10n

(cd LocalizationTools &&swift run LocalizationTools --import --project-path ../giellakbd-ios/GiellaKeyboard.xcodeproj --l10n-project-path ../giellakbd-ios/locales)

* Split keyboard not supported message into two strings

* Consolidate strings into single Localizable.strings

These were previously separated by target; one for HostingApp and one for Keyboard. This doesn't play nice with Pontoon and doesn't add much for the added complexity.

* Update podfile.lock version

* Revert "Consolidate strings into single Localizable.strings"

This reverts commit 8575a7b647513445887256e692ea0bb1a0085bbc.

* I hate that this is the fix for duplicate strings upon export to Pontoon

Turns out exporting strings will always add any discovered NSLocalizedStrings in code to the BaseKeyboard lproj, even if those strings are commented out in code, and even if a BaseKeyboard lproj doesn't exist. Since the Localizable.strings file is added to the Keyboard target, this means the export finds these strings twice: once for Keyboard and once for BaseKeyboard. The workaround is to add Localizable.strings to the BaseKeyboard target as well - then it considers the strings to only be a part of Keyboard, which results in no more duplicate strings showing up in Pontoon.

* Try again: consolidate strings into single Localizable.strings

* Revert "Try again: consolidate strings into single Localizable.strings"

This reverts commit c9820c4bd1734d96b44e21b8d8d2a9682dfea469.

* Reapply "Try again: consolidate strings into single Localizable.strings"

This reverts commit 52ce46baec51139ae9931d6ec702ca392ba9fac0.

* Revert "Get building in Xcode 16.2"

This reverts commit 9b4ade2962cc1d50fff864b02a477d72e61a3272.

* Add test workflow

* Do push-based sample gh-action instead

* First attempt at actual localization imports action

* Rename and try a push trigger

* Add temp logging

* Less verbose

* Attempt to fix path, remove logging

* Try only committing .lproj strings

* Update restore in case we ever need it

* Ignore import strings log